Do not use dnfpluginscore.logger
dnfpluginscore is not core support for dnf plugins, it is just a
collection of "core plugins" for dnf.  There is equally
dnfpluginsextras and dnfpluginsextras.logger.

A comment in dnf.logger comment says the logger name 'dnf.plugin' is "api".
So we can safely rely on that name.

Technically you can install etckeeper without the dnfpluginscore package
(at least I managed to run into this for python2, on a Fedora 29 system
which uses python3 for dnf by default).